requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.35.7
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
